chore: migrate every components to VUE
