+++ apps/api/resources/js/Components/Table/TableContainer.vue
@@ -1,0 +1,16 @@
+<script setup lang="ts">
+import { css } from '@style/css';
+</script>
+
+<template>
+  <table
+    :class="
+      css({
+        borderCollapse: 'separate',
+        borderSpacing: 0,
+      })
+    "
+  >
+    <slot />
+  </table>
+</template>
