+++ apps/api/resources/js/Components/Table/TableRoot.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+import { css } from '@style/css';
+</script>
+
+<template>
+  <div :class="css({ overflowX: 'auto' })">
+    <slot />
+  </div>
+</template>
