--- apps/api/resources/js/Components/TextInput.tsx
@@ -1,13 +1,0 @@
-import { styled } from '@style/jsx';
-
-export const TextInput = styled('input', {
-  base: {
-    textStyle: 'body',
-    borderWidth: 2,
-    borderStyle: 'solid',
-    borderColor: 'green.950',
-    borderRadius: 'md',
-    px: 2,
-    py: 1,
-  },
-});
