--- apps/api/resources/js/Components/index.ts
@@ -1,9 +1,0 @@
-export * from './Button';
-export * from './DefinitionList';
-export * from './FieldContainer';
-export * from './Message';
-export * from './Navbar';
-export * from './Pagination';
-export * from './Table';
-export * from './TextInput';
-export * from './Typography';
