--- apps/api/resources/js/Pages/Layout.tsx
@@ -1,38 +1,0 @@
-import React, { PropsWithChildren } from 'react';
-
-import { Navbar } from '~/Components';
-import { css } from '@style/css';
-import { usePage } from '@inertiajs/inertia-react';
-
-type LayoutProps = PropsWithChildren<{}>;
-
-export function Layout({ children }: LayoutProps) {
-  const { props } = usePage();
-  const isLoggedIn = props.user !== null;
-
-  return (
-    <>
-      {isLoggedIn && <Navbar />}
-      <div
-        scroll-region="true"
-        className={css({
-          w: 'full',
-          h: 'full',
-          overflowY: 'auto',
-        })}
-      >
-        <main
-          className={css({
-            p: '4',
-            px: '8',
-            width: 'full',
-            minHeight: 'full',
-            mx: 'auto',
-          })}
-        >
-          {children}
-        </main>
-      </div>
-    </>
-  );
-}
